+++ crates/full-node/sov-sequencer/src/preferred/mod.rs
@@ -198,7 +198,7 @@ where
             } else {
                 Box::new(RocksDbBackend::new(storage_path).await?)
             };
-        let (db, latest_db_event_id, next_sequence_number, db_cache) =
+        let (db, _latest_db_event_id, next_sequence_number, db_cache) =
             PreferredSequencerDb::<S, Rt>::new(
                 db_backend,
                 shutdown_sender.clone(),
@@ -275,6 +275,8 @@ where
             startup_transaction_cache_writer: None, // The main executor must *not* write to the tx cache. That's handled by the side effects task
         };
 
+        let (replica_sync_notifer, replica_sync_receiver) = watch::channel(None);
+
         let tx_queue_id = Arc::new(AtomicU64::new(0));
         let (synchronized_state, synchronized_state_updator) = create(
             latest_state_update.clone(),
@@ -287,6 +289,7 @@ where
             executor_events_sender,
             next_sequence_number,
             in_flight_blobs,
+            replica_sync_notifer,
             stop_at_rollup_height,
         );
 
@@ -333,8 +336,7 @@ where
                 spawn_replica_sync_task(
                     seq.clone(),
                     shutdown_receiver.clone(),
-                    latest_state_update.clone(),
-                    latest_db_event_id,
+                    replica_sync_receiver,
                 )
                 .await,
             );
@@ -667,6 +669,10 @@ pub(crate) enum PreferredSeqOperation {
     // This is by far the most common scenario, i.e. a nominal
     // `update_state` call during sequencer execution with no unusual
     // conditions.
+    // (is_startup_or_resync, time_spent_fetching_batches)
+    // * is_startup_or_resync: true if the node just started, or just finished resyncing. This is
+    //   needed to update caches (i.e. the transaction_cache).
+    // * time_spent_fetching_batches: for metrics.
     ReplaySoftConfirmationsOnTopOfNodeState(bool, Duration),
 }
 
@@ -746,13 +752,13 @@ where
         }
 
         PreferredSeqOperation::ReplaySoftConfirmationsOnTopOfNodeState(
-            is_startup,
+            is_startup_or_resync,
             time_spent_fetching_batches,
         ) => {
             seq.replay_soft_confirmations_on_top_of_node_state(
                 info,
                 timer_start,
-                is_startup,
+                is_startup_or_resync,
                 time_spent_fetching_batches,
             )
             .await?;